+++ src/controls/taxonomyPicker/TermPicker.tsx
@@ -6,6 +6,8 @@ import styles from './TaxonomyPicker.module.scss';
 import { ITaxonomyPickerProps } from './ITaxonomyPicker';
 import { IWebPartContext } from '@microsoft/sp-webpart-base';
 import * as strings from 'ControlStrings';
+import { Icon } from 'office-ui-fabric-react';
+import { ApplicationCustomizerContext } from '@microsoft/sp-application-base';
 
 export class TermBasePicker extends BasePicker<IPickerTerm, IBasePickerProps<IPickerTerm>>
 {
@@ -18,7 +20,7 @@ export interface ITermPickerState {
 
 export interface ITermPickerProps {
   termPickerHostProps: ITaxonomyPickerProps;
-  context: IWebPartContext;
+  context: IWebPartContext | ApplicationCustomizerContext;
   disabled: boolean;
   value: IPickerTerms;
   allowMultipleSelections : boolean;
@@ -70,7 +72,7 @@ export default class TermPicker extends React.Component<ITermPickerProps, ITermP
           !term.disabled && (
             <span className={styles.pickedTermCloseIcon}
               onClick={term.onRemoveItem}>
-              <i className="ms-Icon ms-Icon--Cancel" aria-hidden="true"></i>
+              <Icon iconName="Cancel" />
             </span>
           )
         }